+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Extract `AbstractProcessorTest.processors()` to simplify getting complete set of processors in downstream projects.
+* Extract `AbstractProcessorTest.compiler()` to simplify setup of compiler.
 
 ### [v0.25](https://github.com/realityforge/proton/tree/v0.25) (2020-01-23) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.24...v0.25)